2002-11-15 Martin Weber <martweb@gmx.net>
	* autotrace.c, fit.c, input-tga.c,
	output-cgm.c, output-dr2d.c, output-dxf.c,
	output-fig.c, outputp2e.c, pxl-outline.c:
	Fixed compilation errors and warnings for certain C compilers.

	* output-pstoedit.h: Fixed compilation errors with Visual C++.

	* output.c: Removed unnecessary streq.

+++ input-tga.c
@@ -279,7 +279,7 @@ ReadImage (FILE              *fp,
   unsigned char *buffer;
   unsigned char *alphas;
 
-  int width, height, bpp, abpp, pbpp, nalphas;
+  unsigned short width, height, bpp, abpp, pbpp, nalphas;
   int j, k;
   int pelbytes, wbytes, bsize, npels, pels;
   int rle, badread;
